+++ AnnounceMoonAndWeatherChange/AnnounceMoonAndWeatherChange.cs
@@ -1,4 +1,6 @@
-﻿using BepInEx;
+﻿using System;
+using AnnounceMoonAndWeatherChange.WeatherWarningAnimations.Impl;
+using BepInEx;
 using BepInEx.Logging;
 using HarmonyLib;
 
@@ -7,20 +9,24 @@ namespace AnnounceMoonAndWeatherChange;
 [BepInPlugin(MyPluginInfo.PLUGIN_GUID, MyPluginInfo.PLUGIN_NAME, MyPluginInfo.PLUGIN_VERSION)]
 [BepInDependency("WeatherTweaks", BepInDependency.DependencyFlags.SoftDependency)]
 public class AnnounceMoonAndWeatherChange : BaseUnityPlugin {
+    internal static ConfigManager? configManager;
+
+    // ReSharper disable once FieldCanBeMadeReadOnly.Global
+    // ReSharper disable once MemberCanBePrivate.Global
+    public static EventHandler registerNowEvent = null!;
     public static AnnounceMoonAndWeatherChange Instance { get; private set; } = null!;
     internal new static ManualLogSource Logger { get; private set; } = null!;
     private static Harmony? Harmony { get; set; }
-    internal static ConfigManager? configManager;
 
     private void Awake() {
         Logger = base.Logger;
         Instance = this;
 
+        InitializeDefaultAnimations();
+
+        registerNowEvent?.Invoke(this, null);
+
         Patch();
-
-        configManager = new ConfigManager(Config);
-
-        configManager.HandleConfig();
 
         if (DependencyChecker.IsWeatherTweaksInstalled()) {
             WeatherTweaksSupport.enabled = true;
@@ -30,8 +36,14 @@ public class AnnounceMoonAndWeatherChange : BaseUnityPlugin {
         Logger.LogInfo($"{MyPluginInfo.PLUGIN_GUID} v{MyPluginInfo.PLUGIN_VERSION} has loaded!");
     }
 
+    private void Start() {
+        configManager = new(Config);
+
+        configManager.HandleConfig();
+    }
+
     internal static void Patch() {
-        Harmony ??= new Harmony(MyPluginInfo.PLUGIN_GUID);
+        Harmony ??= new(MyPluginInfo.PLUGIN_GUID);
 
         Logger.LogDebug("Patching...");
 
@@ -46,5 +58,11 @@ public class AnnounceMoonAndWeatherChange : BaseUnityPlugin {
         Harmony?.UnpatchSelf();
 
         Logger.LogDebug("Finished unpatching!");
+    }
+
+    private static void InitializeDefaultAnimations() {
+        SlidingAnimation.Initialize();
+        FadeAnimation.Initialize();
+        BlinkingAnimation.Initialize();
     }
 }