+++ AnnounceMoonAndWeatherChange/DependencyChecker.cs
@@ -4,7 +4,6 @@ using BepInEx.Bootstrap;
 namespace AnnounceMoonAndWeatherChange;
 
 internal static class DependencyChecker {
-    internal static bool IsWeatherTweaksInstalled() {
-        return Chainloader.PluginInfos.Values.Any(metadata => metadata.Metadata.GUID.Contains("WeatherTweaks"));
-    }
+    internal static bool IsWeatherTweaksInstalled() =>
+        Chainloader.PluginInfos.Values.Any(metadata => metadata.Metadata.GUID.Contains("WeatherTweaks"));
 }