+++ AnnounceMoonAndWeatherChange/CHANGELOG.md
@@ -1,6 +1,13 @@
 # Announce Moon And Weather Change Mod<br>
 
+## 1.2.0 (All the animations Update)
+
+\+ Added 2 more animations (Outside mods can technically also add new animations)<br>
+\# Big Code restructure<br>
+\# Please note that your old config will be wiped!<br>
+
 ## 1.1.0 (Sliding Update)<br>
+
 \+ Added sliding animation to weather warning<br>
 \+ Added more config options (Font size, color, ...)<br>
 \# Weather warning background now resizes with text<br>
@@ -8,4 +15,5 @@
 \# Project is now open source<br>
 
 ## 1.0.0
+
 \+ Initial Release<br>